pyproject.toml
[project]
name = "memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.3.3",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "loguru>=0.7.3",
    "pydantic-ai>=0.1.11",
    "python-dotenv>=1.1.0",
]
